+++ rules/legacy.js
@@ -1,11 +1,11 @@
 // Legacy
 module.exports = {
   'rules': {
-    'max-depth': [0, 4],
-    'max-len': [0, 80, 4],
-    'max-params': [0, 3],
-    'max-statements': [0, 10],
-    'no-bitwise': 2,
-    'no-plusplus': 0
+    'max-depth': 'off',
+    'max-len': 'off',
+    'max-params': 'off',
+    'max-statements': 'off',
+    'no-bitwise': 'error',
+    'no-plusplus': 'off'
   }
 };