+++ rules/strict.js
@@ -1,6 +1,6 @@
 // Strict
 module.exports = {
   'rules': {
-    'strict': 0
+    'strict': 'off'
   }
 };